refactor: delete rl/config.py, inline Config into async_rl_loop.py
Config is RL-facing — users should see it in the loop file alongside
reward, filter, and training logic. Now async_rl_loop.py is fully
self-contained: Config, reward, filter, datum, scheduler, and loop.
